[DEL] deleted OnboardingViewFour
